Extracting class to build events search filters.
